scripts/requirements.txt
spacy
tweepy
farcaster
python-dotenv
transformers
torch